[Search] Set up CI: lint, format, test, security (#86)
# Description
<!-- Please include a summary of the change(s) and which issue is being
fixed. Please provide as much detail as possible. -->
Create the base CI jobs for the search service: linting, formatting,
testing, and security

This PR is stacked on top of #85, so first review that PR (as this diff
contains that PR diff also).

<!-- Replace `XXX` with the concerning issue number. The \"#\" links
this PR to its relevant issue -->
Closes #22 

## How to Test
<!-- Provide some simple steps so that others can verify your
implementation -->
View the CI output below or run the actions locally with
[act](https://github.com/nektos/act)

## Checklist
- [x] The code includes tests if relevant
- [x] I have *actually* self-reviewed my changes and done QA