Sejal/frontend121 create plain pages for martletplace (#123)
# Description
<!-- Please include a summary of the change(s) and which issue is being
fixed. Please provide as much detail as possible. -->
Added separate pages for application. 

This PR is a quick and dirty template for the pages we'll need. We will
need to add protected routes for the pages that require authentication.
We may also need to have a `listingId`, `userId`, etc when editing or viewing a
listing.


<!-- Replace `XXX` with the concerning issue number. The "#" links this
PR to its relevant issue -->
Closes #121 

## How to Test
<!-- Provide some simple steps so that others can verify your
implementation -->


## Checklist
- [x] The code includes tests if relevant
- [x] I have *actually* self-reviewed my changes and done QA
<!-- Only check this off if you have actually done a self-review! DO NOT
request any review from others until you have done your self-review! -->

---------

Co-authored-by: Sejal Agarwal <sejalagarwal@Sejals-MacBook-Air.local>
Co-authored-by: Sejal Agarwal <sejalagarwal@w134-87-173-149.wireless.uvic.ca>