[Messaging] Mock the functionality of messaging (#252)
# Description
<!-- Please include a summary of the change(s) and which issue is being
fixed. Please provide as much detail as possible. -->
This is the backup for messaging, if it doesn't get finished in time,
this will replace areas that call for messaging with a mailto link

<!-- Replace `XXX` with the concerning issue number. The "#" links this
PR to its relevant issue -->
Closes #75 

## How to Test
<!-- Provide some simple steps so that others can verify your
implementation -->
Run the app

## Checklist
- [ ] The code includes tests if relevant
- [ ] I have *actually* self-reviewed my changes and done QA
<!-- Only check this off if you have actually done a self-review! DO NOT
request any review from others until you have done your self-review! -->

---------

Co-authored-by: Max Thomson <git@mnt.dev>