Fix flaky recommender test (#419)
# Description

The recommender tests were flaky after an attempted fix. I believe that
fix didn't work since, when we retry loading the model when it gets
corrupted, we don't remove the models first. This PR cleans up the code
surrounding this issue to make it more transparent, and removes the
models before redownloading the model.

Closes #310 

## Checklist
- [x] The code includes tests if relevant
- [x] I have *actually* self-reviewed my changes and done QA