Create the Create Charities Page (#402)
# Description
This PR adds the route, and page, and mocks the endpoints for creating a
listing. This page is for testing charities and isn't intended for
regular users(unless we're otherwise told). Therefore, I've kept things
like console logs and added buttons so that the state of the charity
object can be known by whoever is using this page.


![image](https://github.com/user-attachments/assets/fdd5fcbc-75f0-4ae5-97e0-0b7ec377c865)

![image](https://github.com/user-attachments/assets/dbc73b48-dcfd-4e32-9553-4602cdcf894e)


![image](https://github.com/user-attachments/assets/5ae41757-62fc-4a17-9dfd-0878bbb55207)

Closes #304 

## How to Test
Go to [localhost/charity/new](http://localhost/charity/new), or run `npm
test` or run `npx cypress open` from within the frontend directory

## Checklist
- [x] The code includes tests if relevant
- [ ] I have *actually* self-reviewed my changes and done QA