[Search] Run SQL in background to 3x reponse time (#421)
# Description
The SQL INSERT does not need to happen while the client is waiting. The
query now executes in a [FastAPI background
task](https://fastapi.tiangolo.com/tutorial/background-tasks/). Any
errors resulting from the query will not cause the search to fail but is
not concerning as missing the data for a single search history is
outweighed by the performance gains.

closes: #422

## How to Test
| Before| After|
|---|---|

|![image](https://github.com/user-attachments/assets/b6a782ad-cd1f-4b60-93ed-21e530aadc42)|![image](https://github.com/user-attachments/assets/2c71ae91-234a-4df3-810d-dc9dcd667773)|

## Checklist
- [ ] The code includes tests if relevant
- [x] I have *actually* self-reviewed my changes and done QA