Gavin/348 update search endpoint to return charity status (#521)
# Description
Updated the search endpoint to return the CharityID field in accordance
with the API contract:

![image](https://github.com/user-attachments/assets/5def676c-f86c-4764-bdd8-06c9a0edae56)

On the frontend it looks like this. Nice job Chloe!

![image](https://github.com/user-attachments/assets/e872f233-e775-4787-87e9-8b372d3c7c96)

Most of the changes are updating test_server.py to use the new charity
endpoint.

There are a few frontend changes which are as follows:
- Changing charityId to CharityID to comply with the search and
recommender endpoints
- Making CharityID nullable (since null is used to check whether an item
is NOT for charity)

There is still this issue on the frontend, which I suspect is because
the recommender does not yet have returned charityID values. If someone
knows its for a different reason please let me know and I'll change it
promptly:

![image](https://github.com/user-attachments/assets/82f9af8a-4071-404a-9949-5004cd0ba99f)


Closes #348

## How to Test
* Build the application
* Create an account/login
* Search for some stuff (**some** items in search should be labeled as
for charity like so):

![image](https://github.com/user-attachments/assets/8a1e11af-9390-4484-9c28-d31fef3c8c4a)

* Once the app is built, run these two commands in a separate terminal:
`docker exec -it martletplace_search bash` then `pytest
src/test_server.py`


## Checklist
- [X] The code includes tests if relevant
- [X] I have *actually* self-reviewed my changes and done QA